Fresco Vito: Added debug overlay
Summary:
This adds a simplified debug overlay for Fresco Vito.
Previously, most of the logic lived in the drawable itself. However, in order to have a more flexible debug overlay with potentially different configurations, it's better to have a dumb drawable and an interface to set it up however we'd like.

Right now, it displays the URI and adds an orange overlay color.

Reviewed By: cgrushko

Differential Revision: D15555432

fbshipit-source-id: 419e36fb42ef0b641533e8d352e907d85fda7542