Animated drawable2 bugfix (#2355)
Summary:
## Motivation

As #2205 comments, sometimes AnimatedDrawable2 finish playing animation without calling onAnimationStop.
I read the source code and find out that in AnimatedDrawable2.java line 221-230:
```
long actualRenderTimeEnd = now();
if (mIsRunning) {
      // Schedule the next frame if needed.
    targetRenderTimeForNextFrameMs =
          mFrameScheduler.getTargetRenderTimeForNextFrameMs(actualRenderTimeEnd - mStartTimeMs);
    if (targetRenderTimeForNextFrameMs != FrameScheduler.NO_NEXT_TARGET_RENDER_TIME) {
         scheduledRenderTimeForNextFrameMs =
            targetRenderTimeForNextFrameMs + mFrameSchedulingDelayMs;
         scheduleNextFrame(scheduledRenderTimeForNextFrameMs);
     }
}
```
If animation still running but actualRenderTimeEnd - mStartTimeMs is larger than loopDurationMs, there will not be scheduleNextFrame, so onAnimationStop will not be calling.
To fix this bug, I added few lines in the 'else' branch:
```
    long actualRenderTimeEnd = now();
    if (mIsRunning) {
      // Schedule the next frame if needed.
      targetRenderTimeForNextFrameMs =
          mFrameScheduler.getTargetRenderTimeForNextFrameMs(actualRenderTimeEnd - mStartTimeMs);
      if (targetRenderTimeForNextFrameMs != FrameScheduler.NO_NEXT_TARGET_RENDER_TIME) {
        scheduledRenderTimeForNextFrameMs =
            targetRenderTimeForNextFrameMs + mFrameSchedulingDelayMs;
        scheduleNextFrame(scheduledRenderTimeForNextFrameMs);
      }else{
        mAnimationListener.onAnimationStop(this);
        mIsRunning = false;
      }

```
If there is no next frame, the animation is finished, so onAnimationStop should be called and mIsRunning should be false.
## Test Plan (required)

Unfortunately, we can not reproduce the problem every time, but I assume that when the frame interval is vary small such as 1ms, this problem can be reproduced easily.
Pull Request resolved: https://github.com/facebook/fresco/pull/2355

Reviewed By: defHLT

Differential Revision: D15693521

Pulled By: oprisnik

fbshipit-source-id: b7a01d2eafaa35a7e5c8b43bac0e5e7ce95e7f09