Adds a color filter and placeholder options to FrescoVitoLithoImageOptionsConfigFragment
Summary:
Allows the user to darken the image with a given tint color, or to leave the image unchanged.

Adds a further set of options for defining the placeholder that is shown while an image loads. Available options are: none, image (the Fresco logo is used here) and a block color (red in this case).

Reviewed By: oprisnik

Differential Revision: D16494294

fbshipit-source-id: 03076d8d1b98f9ae3ae3a119f41d4477026e6be0