Use producer listener 2 in producers
Summary:
NOTE: This is a breaking API change if you are using custom producers. If that is the case, please update your producers to use the new API accordingly. You can take a look at this change to see how we did this for existing producers.

We now use the new producer listener in all producers and tests. The internal listener will still delegate to the old ProducerListener, which is conform with the current API.

Reviewed By: defHLT

Differential Revision: D16827972

fbshipit-source-id: 747069061a7007e92abfbf66da7c7e8e268fc50e