Added experiment to move image fetch to lightweight BG thread
Summary: This experiment improves the mount time for images by moving work from the UI thread to a lightweight background thread. On the UI thread, we only check if the image is in cache. If not, we post a runnable to the background thread to actually kick off the image request. The runnable will check if the image is still attached (i.e. still visible) and kick off work if needed.

Reviewed By: shrisukhani

Differential Revision: D18247803

fbshipit-source-id: 0f8baa042d8d8b72f603602ceb0b0ddd55c98511