Move from FrescoContext to FrescoContextProvider
Summary:
We shouldn't need to expose FrescoContext. Instead, the FrescoContextProvider interface is used to be able to retrieve the main components for Vito UI code.
This change also simplifies the provider API, we don't need the Context or Caller Context when we get a Vito component since they are managed solely by the provider and should not depend on image-specific data.

Reviewed By: defHLT

Differential Revision: D20036984

fbshipit-source-id: c43ada2c46ea3863fce556f2b2551651e89e045a