Remove Drawee-Volley module
Summary:
NOTE: If you're using Drawee-Volley, you can still continue using it by just copying the files removed here into your project. We did not significantly change these in a long time, so this has been stable.

The Drawee-Volley Maven artifact is only downloaded by a very small number of users and we are not actively improving Drawee-Volley for years.

Reviewed By: defHLT

Differential Revision: D20646885

fbshipit-source-id: 9fdb20b55f4b2ea8ace07cf1765667ed08648672