Vito2: set fetch submitted if image comes from cache shortcut
Summary:
Vito2 will only submit the image fetch once. If you re-mount the component, we can detect if it's already displayed.
However, if the image comes from the cache shortcut, we don't have to submit a fetch, but should still mark that we "found" the image. Otherwise, subsequent mounts will re-trigger instead of being a no-op. They should still come from the cache but it will still be an unnecessary overhead.

Reviewed By: muraziz

Differential Revision: D21645338

fbshipit-source-id: 9efa6fd6e0e34f36afe292ca0a81378fff062631