Don't use CallerContext for VitoImageRequest CachedValue creation
Summary:
The CachedValue API is using a HashMap with all props as input, where it will use `hashCode()` for each prop.
This means that for different CallerContext objects, the VitoImageRequest might not be cached properly. We should keep the VitoImageRequest independent of the caller context since we do have that information when the image is mounted, where we set it for the MountContent which forwards all requests to loggers etc.

Reviewed By: defHLT

Differential Revision: D21903865

fbshipit-source-id: 025169ffe9c3f65d84390accaad551e81fad78ff