Move VitoDrawableFactory to vito-options and add API to override factory via ImageOptions
Summary:
It should be possible to set a custom drawable factory for a given image via ImageOptions.
In order for this to work, we need to move the factory to the ImageOptions package to avoid cyclic dependencies and to have API components in a consistent location.
It also makes sense to rename the drawable factory since it creates the drawable based on the image and image options.

Reviewed By: colriot

Differential Revision: D23216704

fbshipit-source-id: ae01e01a91813c7b2d6377a7adb801cbd4547372