There is a period of time between checking if NativeLoader has been initialized, and when we try to perform .init() which means a race condition can occur (as mentioned in https://github.com/facebook/SoLoader/issues/60).
This change uses synchronization on the NativeLoader class, and a secondary check to verify that the initialization needs to take place just before we perform it, in order to reduce the window for a race condition to affect client code.

When combined with https://github.com/facebook/SoLoader/pull/69 this will remove the race condition from the code because the initialisation inside SoLoader and inside Fresco will both be synchronized on the NativeLoad.class object.