Update SoLoader to v0.10.0 (#2581)
Summary:
Thanks for submitting a PR! Please read these instructions carefully:

- [X] Explain the **motivation** for making this change.
- [X] Provide a **test plan** demonstrating that the code is solid.
- [X] Match the **code formatting** of the rest of the codebase.
- [X] Target the `master` branch

## Motivation (required)

Update to PR#2579 to make use of the new API method exposed by SoLoader 0.10.0. Using the new API means that the responsibility of ensuring the SoLoader NativeLoader initialisation checks and initialisation moves to SoLoader where it can be updated as appropriate and all SoLoader clients will benefit.

PR#2579 required other clients to be aware of the class we were synchronizing on, whereas with SoLoaders API that requirement no longer exists.

## Test Plan (required)

There are no functional changes in this PR, but there is a library update. Existing tests will ensure that the exposed behaviours of Fresco aren't changed.

Pull Request resolved: https://github.com/facebook/fresco/pull/2581

Reviewed By: wizh

Differential Revision: D26173493

Pulled By: oprisnik

fbshipit-source-id: 05e2846b98273dc3df43b5d6f54f44a3c19761e2