Update LocalVideoThumbnailProducer.java
Thumbnail.createVideoThumbnail fails in Android 10, as such implementation uses the File API which in API 29 cannot be used to access public media. For package private media is still valid. To solve the issue, when createVideoThumbnail fails, fallback to createThumbnailFromContentProvider, as such method uses the media Uri, which works on all API levels.