Live editing: Use Vito image perf listener instead of Drawable listener for more accurate on-screen image tracking
Summary:
We can use image perf listener to track which images are mounted / unmounted, which means that the selector is now much more accurate since it doesn't just track any Drawable forever but actually keeps a pool of on-screen images to be edited.
This also means we don't have to keep the weak references around for too long and allows us to also remove the Drawable listener altogether, simplifying the overall setup.

In addition, images should always be in mount order, which means that they are not in a random order as would have been possible before.

Reviewed By: defHLT

Differential Revision: D32645080

fbshipit-source-id: 0086daf67b286d2e355ba8ea0d085c2234ec67eb