Fall back to static rendering for animated WebPs instead of throwing
Summary:
For animated GIFs, we properly fall back to static decoding if no dedicated animated GIF decoder is added.
However, for WebP, we simply throw an exception and no image shows up, even when Android can display the first frame (API30).

With this diff, we change the behavior to be the same for animated WebPs, where a static image will be displayed if supported by the system.

Differential Revision: D36512321

fbshipit-source-id: 72db7bc0466a279d41c589eae450a56cc8ede357