Change ImageSource architecture
Summary:
With this change, ImageSource has an improved architecture, especially when not used with the Fresco ImagePipeline. Previously, the ImageSource was tied to ImagePipeline concepts (`DataSource`, `ImageRequest`) and had logic to work with these, such as extracting the "final image request".

Now, there is a new adapter that converts the simplified ImageSource (e.g. one that just holds a URI) to whatever the underlying image pipeline needs. For example, for the Fresco ImagePipeline, it would help with setting up the DataSource for the image fetch.
It also means that we do not have to initialize the ImageSourceProvider since it can now directly create these simplified objects, without needing to know about the underlying Vito implementation.

Since displaying an image based on an ImagePipeline-specific ImageRequest is fairly common, there is a new `ImagePipelineImageSource` that allows to create `ImageRequest`s - and that only works with the Fresco ImagePipeline. Hence, it still holds some logic specific for the Fresco ImagePipeline (fetch based on the `ImageRequest)`

Reviewed By: defHLT

Differential Revision: D34420550

fbshipit-source-id: 45a622efc3b87ea66fe6663e6c47e6db581cfb98