Add Resources parameter to ImageOptionsDrawableFactory interface to simplify initialization and ensure proper Resource support
Summary:
By supplying Resources directly when a Drawable is created, we eliminate the need for some Drawable factories (e.g. the Bitmap Drawable factory) to hold on to Resources, which have to be supplied when Vito is initialized. As a result, Vito's initialization now doesn't require the Resources parameter any more, which aids in simplifying the initialization logic in the future.

It also means that we can properly use themed Resources that might be set for a given View.

Reviewed By: defHLT

Differential Revision: D37244690

fbshipit-source-id: c65ed7a0149e41e00ba02c54e65b7cef9e37f315