Implement Drawable.Callback for KFrescoVitoDrawable to fix layer callbacks
Summary:
KFrescoVitoDrawable would pass its Drawable.Callback to layers. If the layer would for example be a Drawable and the Callback is an ImageView, the Drawable could notify the ImageView that it needs to invalidate - but the ImageView doesn't know about the Drawable, it only knows about KFrescoVitoDrawable.
This change implements Drawable.Callback for KFrescoVitoDrawable, which will then instruct its Callback to properly update.
This is aligned with how this works for the Java implementation / ArrayDrawable as well.

Reviewed By: steelrooter

Differential Revision: D43270935

fbshipit-source-id: 75dc18542830f3122139f15d02bdf26950cf42f1