Added Vito examples for overlay rounding
Summary:
Drawee offers an "overlay color" rounding mode, which draws a solid color outside the image so that it appears rounded.

For Vito, we did not add a separate rounding mode for this. However, it's easy to set up using an overlay with a  `RoundedCornersDrawable` where an outer color is specified. This is aligned with how Drawee would achieve this under the hood.

Reviewed By: defHLT

Differential Revision: D46222141

fbshipit-source-id: 12ce3340555fee684d4c5c6197b73d74cbe6c092