Added a reason string to EmptyImageSource to be able to distinguish between different cases
Summary:
This adds a new `reason` to `EmptyImageSource` which can be useful to indicate why a given image is empty (e.g. if it's expected or if a data model returned a `null` image URI or of something else went wrong).

If it is expected for the image to be empty, developers can use `ImageSourceProvider.emptySource()` and otherwise create a custom `EmptyImageSource("reason why the image is empty")`.

Reviewed By: kartavya-ramnani

Differential Revision: D80084964

fbshipit-source-id: 7cda71dc17f7314f59f2a4dba455a40cde8847c0